requirements.txt
python-dotenv
langchain
langchain-google-genai
gTTS
streamlit 